requirements.txt
discord.py[voice]
datetime
requests
youtube-dl
gTTS
urllib3
feedparser
disputils

fortnite-api

asyncio
aiohttp
aiofile